# openapi.yaml
openapi: 3.1.0
info:
  title: komondor-api
  version: "1.0.0"
  summary: TSL Sequence Data API — projects, samples, sequencing runs and their files.
  description: |
    A description of what komondor-api **does today**, derived by reading
    `routes/` and `models/`. It is a record of current behaviour, not a
    proposal. Where the current behaviour is a known defect it is documented as
    it is, and the operation carries a `DRIFT` or `DEFECT` note saying so.
    Fixing one of those is a change to the API and must update this file in the
    same commit.

    `__tests__/contract/openapi.test.js` fails if a route exists in `routes/`
    with no entry here, so the two cannot silently diverge.

    ## Known drift from the sibling clients

    Three repositories consume this API: `komondor-web` (Nuxt 2 UI),
    `komondor-power` (bulk CSV ingest) and `komondor-nudge` (a scheduled
    worker that reads the same MongoDB directly rather than over HTTP). The
    following are the places where their assumptions and this server's
    behaviour have come apart. See `docs/CONTRACTS.md` for the policy.

    1. **Routes are mounted at the ROOT, not under `/api`.** `app.js` calls
       `app.use(authRoutes)` with no path prefix, so login is `POST /login`,
       not `POST /api/login` or `POST /auth/login`. The comment above the login
       handler in `routes/auth.js` still reads "Add POST - /api/login", which
       is where the belief came from. `komondor-power`'s client tries
...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /accessions/new:
    post:
      tags: [Accessions]
      summary: Attach ENA accessions to a project, sample or run.
      description: |
        Requires the `hasFullRecordsAccess` population — admins, the literal
        `admin` account, and usernames listed in `FULL_RECORDS_ACCESS_USERS` —
        NOT mere group membership. This is the return leg of the
        `/accessions/csv` export that the same predicate already gates: an
        accession is the identifier ENA issues, not the group's own metadata,
        and its consequences are external and public. Group write access would
        also have been the wrong gate, since an ENA admin is typically not a
        member of the groups they submit for.

        The record's group is still checked, as a READ: the group must be live.
        Because `Group.GroupsIAmIn` excludes soft-deleted groups for everybody,
        a retired group stops authorising accession writes into records nobody
        can see any more, and a record with a dangling group reference is
        refused rather than written.

        `accessions` is REQUIRED and REPLACES the stored array; it does not
        append. Omitting it used to write `undefined`, which mongoose strips —
        the request answered 200 having changed nothing. That silent no-op is
        now a 400. "Clear the accessions" is an explicit `[]`.

        `releaseDate` is only honoured when `type` is `project`.

        On success the body is EMPTY — `res.status(200).send()` with no
        argument. Do not parse it as JSON.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, typeId, accessions]
              properties:
                type:
                  type: string
                  enum: [project, sample, run]
                typeId:
                  $ref: "#/components/schemas/ObjectId"
                accessions:
                  type: array
                  description: |
                    Required. Replaces the stored array wholesale. Send `[]` to
                    clear.
                  items:
                    type: string
                releaseDate:
                  type: string
                  description: "DD-MM-YYYY. Stored as a string; ignored unless `type` is `project`."
      responses:
        "200":
          description: Updated. Empty body.
        "400":
          description: |
            Bad `type`, missing or malformed `typeId`, or `accessions` absent
            or not an array of strings.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/GroupPermissionDenied"
        "404":
          description: No record of that type with that id.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Update failed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /accessions/csv:
    get:
      tags: [Accessions]
      summary: Cross-group CSV export of every run. Restricted.
      description: |
        Ignores group membership by design — it returns every run in the
        database — so it is gated on `hasFullRecordsAccess`: admins, the
        literal `admin` account, and usernames listed in
        `FULL_RECORDS_ACCESS_USERS`. NOT `isAdmin` alone; the ENA admins who
        use this do not carry that claim. See BREAKING_CHANGES §13.

        The CSV is returned as a STRING inside a JSON object, not as a
        `text/csv` body. The heading row keeps a trailing comma, which
        consuming services parse. Fields containing a delimiter are quoted
        (BREAKING_CHANGES §10).

        Rows whose sample, project or group cannot be resolved are skipped and
        logged, so the export can be shorter than the run count.

        `list_of_read_files` contains the effective authoritative location:
        the configured HPC display root for `hpc`/`migrating` projects and
        `s3://` URIs for verified AWS copies. An inconsistent historical file
        path is emitted as `unresolved:<stored path>` rather than pointing at a
        nonexistent HPC file or duplicating the project prefix. One export may
        therefore contain both POSIX and S3 locations.

        DRIFT: `FULL_RECORDS_ACCESS_USERS` on this server must be kept in step
        with `ENA_ADMINS` in komondor-web, which decides who is offered the
        export page. Two hand-maintained allowlists in two repos — this is
        precisely the pattern `docs/CONTRACTS.md` argues against.
      responses:
        "200":
          description: The export.
          content:
            application/json:
              schema:
                type: object
                properties:
                  csv:
                    type: string
                    description: |
                      Columns: group, owner, ena_project_submission_date,
                      project_name, project_id, project_accession, sample_name,
                      sample_id, sample_accession, run_name, run_id,
                      run_accession, run_creation_date, list_of_read_files.
                      Multi-valued cells are `;`-joined.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          description: The caller lacks full-records access.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    const: You do not have permission to export all records
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <token>`, signed with `JWT_SECRET` (HS256) by
        `lib/utils/jwtSign.js`.

        **The `exp` claim is mandatory.** `lib/utils/getUserFromRequest.js`
        rejects a structurally valid, correctly signed token that carries no
        `exp` by raising `TokenExpiredError`, which surfaces as 401. This was
        added deliberately (commit "refuse legacy tokens without an expiry
        claim"): tokens issued before August 2026 had no expiry and stayed
        valid forever, including tokens minted during the LDAP incident with an
        empty `groups` claim. Such tokens can only be fixed by re-logging in.

        Tokens expire after `JWT_EXPIRES_IN`, default `7d`.

        **Group membership is a snapshot.** `groups` is resolved once, at
        login, and never re-read. A user added to a group sees nothing new
        until their token expires and they sign in again — that expiry is the
        upper bound on the staleness.

        Requests with NO `Authorization` header are not rejected by the
        app-level middleware; they simply arrive with no `req.user`, and each
        route's `isAuthenticated` guard turns that into a 401. Routes without
        that guard (`/projects/names`, `GET /options/*`, `/health`, `/test`,
        `/login`, `/logout`, `/uploads`) serve anonymous callers.
...
  responses:
    Unauthenticated:
      description: |
        No usable token. Two distinct bodies reach clients as 401:
        `{ error: "Authentication required" }` from a route's
        `isAuthenticated` guard, and
        `{ error: "Invalid or expired authentication token", detail }` from the
        app-level middleware in `app.js` when a token was present but bad or
        expired.
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                properties:
                  error:
                    const: Authentication required
              - type: object
                properties:
                  error:
                    const: Invalid or expired authentication token
                  detail:
                    type: string
                    description: The underlying jsonwebtoken message.
                  requestId:
                    type: string
                    description: Present only when raised by the terminal error handler.
...
    GroupPermissionDenied:
      description: |
        The caller is neither the record's owner, nor a member of its group,
        nor an admin. The message embeds the username:
        `User '<username>' does not have permission to ...`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ServerError:
      description: |
        An unhandled failure. In production a 500 replaces `error` with a
        generic string but STILL includes `detail` — komondor-power is an
        internal client and relies on it for diagnostics.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    ObjectId:
      type: string
      pattern: "^[0-9a-fA-F]{24}$"
      # Deliberately not the old hard-coded '2Blades' group id: that literal
      # used to carry meaning in routes/projects.js and reusing it as a generic
      # example invited the reader to think it still does.
      examples: ["507f1f77bcf86cd799439011"]
    ErrorEnvelope:
      type: object
      description: |
        The standard failure body, produced by `handleError` in
        `routes/_utils.js` and by the terminal handler in `app.js`.

        `detail` carries the underlying error message — more specific than
        `error` (e.g. "E11000 duplicate key error" against "Failed to create
        new project.") — and is present even for production 500s.

        Not universal: `/login` uses `message`; `/groups/*` and `/me` send a
        bare `error`; `/read-file` and `/directory-files` send `{ error }` with
        a 200.
      properties:
        error:
          type: string
          description: A user-facing message.
        detail:
          type: string
          description: The underlying error message, for diagnostics.
        requestId:
          type: string
          description: Correlates with the server log line for this failure.
          examples: ["1735689600000-k3j9x2p1a"]
        code:
          type: string
          description: Optional stable machine-readable error code.
        projectId:
          type: string
          description: Present on project-storage refusals.
        storageState:
          type: string
          description: Present on project-storage refusals; may be `invalid`.
...